# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/budgets": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budget - Budgets"
                ],
                "summary": "List user budgets",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Filter active budgets only",
                        "name": "active_only",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_edalferes_monetics_internal_modules_budget_handler_dto.BudgetResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budget - Budgets"
                ],
                "summary": "Create a new budget",
                "parameters": [
                    {
                        "description": "Budget creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_edalferes_monetics_internal_modules_budget_handler_dto.CreateBudgetRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_edalferes_monetics_internal_modules_budget_handler_dto.BudgetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/categories": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budget - Categories"
                ],
                "summary": "List user categories",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by type (income or expense)",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_edalferes_monetics_internal_modules_budget_handler_dto.CategoryResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budget - Categories"
                ],
                "summary": "Create a new category",
                "parameters": [
                    {
                        "description": "Category creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_edalferes_monetics_internal_modules_budget_handler_dto.CreateCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_edalferes_monetics_internal_modules_budget_handler_dto.CategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/reports/monthly": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budget - Reports"
                ],
                "summary": "Get monthly financial report",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Year",
                        "name": "year",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Month (1-12)",
                        "name": "month",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_edalferes_monetics_internal_modules_budget_handler_dto.MonthlyReportResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/transactions": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budget - Transactions"
                ],
                "summary": "List user transactions",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by account ID",
                        "name": "account_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by category ID",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by type (income, expense, transfer)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by start date (RFC3339)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by end date (RFC3339)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_edalferes_monetics_internal_modules_budget_handler_dto.TransactionResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budget - Transactions"
                ],
                "summary": "Create a new transaction",
                "parameters": [
                    {
                        "description": "Transaction creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_edalferes_monetics_internal_modules_budget_handler_dto.CreateTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_edalferes_monetics_internal_modules_budget_handler_dto.TransactionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_edalferes_monetics_internal_modules_budget_handler_dto.BudgetResponse": {
            "type": "object",
            "properties": {
                "alert_at": {
                    "type": "number"
                },
                "amount": {
                    "type": "number"
                },
                "category_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "is_over_budget": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "percentage_used": {
                    "type": "number"
                },
                "period": {
                    "type": "string"
                },
                "remaining": {
                    "type": "number"
                },
                "should_alert": {
                    "type": "boolean"
                },
                "spent": {
                    "type": "number"
                },
                "start_date": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "github_com_edalferes_monetics_internal_modules_budget_handler_dto.CategoryBreakdownItem": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "category_id": {
                    "type": "integer"
                },
                "category_name": {
                    "type": "string"
                },
                "category_type": {
                    "type": "string"
                },
                "percentage": {
                    "type": "number"
                }
            }
        },
        "github_com_edalferes_monetics_internal_modules_budget_handler_dto.CategoryResponse": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_edalferes_monetics_internal_modules_budget_handler_dto.CreateBudgetRequest": {
            "type": "object",
            "required": [
                "amount",
                "category_id",
                "end_date",
                "name",
                "period",
                "start_date"
            ],
            "properties": {
                "alert_at": {
                    "type": "number"
                },
                "amount": {
                    "type": "number"
                },
                "category_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "period": {
                    "type": "string",
                    "enum": [
                        "monthly",
                        "quarterly",
                        "yearly",
                        "custom"
                    ]
                },
                "start_date": {
                    "type": "string"
                }
            }
        },
        "github_com_edalferes_monetics_internal_modules_budget_handler_dto.CreateCategoryRequest": {
            "type": "object",
            "required": [
                "name",
                "type"
            ],
            "properties": {
                "color": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "income",
                        "expense"
                    ]
                }
            }
        },
        "github_com_edalferes_monetics_internal_modules_budget_handler_dto.CreateTransactionRequest": {
            "type": "object",
            "required": [
                "account_id",
                "amount",
                "category_id",
                "date",
                "description",
                "type"
            ],
            "properties": {
                "account_id": {
                    "type": "integer"
                },
                "amount": {
                    "type": "number"
                },
                "attachments": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "category_id": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "destination_account_id": {
                    "type": "integer"
                },
                "is_recurring": {
                    "type": "boolean"
                },
                "recurrence_end": {
                    "type": "string"
                },
                "recurrence_rule": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "transfer_fee": {
                    "type": "number"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "income",
                        "expense",
                        "transfer"
                    ]
                }
            }
        },
        "github_com_edalferes_monetics_internal_modules_budget_handler_dto.MonthlyReportResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "category_breakdown": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_edalferes_monetics_internal_modules_budget_handler_dto.CategoryBreakdownItem"
                    }
                },
                "month": {
                    "type": "integer"
                },
                "month_name": {
                    "type": "string"
                },
                "total_expense": {
                    "type": "number"
                },
                "total_income": {
                    "type": "number"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "github_com_edalferes_monetics_internal_modules_budget_handler_dto.TransactionResponse": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "integer"
                },
                "amount": {
                    "type": "number"
                },
                "attachments": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "category_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "destination_account_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "is_recurring": {
                    "type": "boolean"
                },
                "parent_id": {
                    "type": "integer"
                },
                "recurrence_end": {
                    "type": "string"
                },
                "recurrence_rule": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "transfer_fee": {
                    "type": "number"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        }
    }
}